feat: move coercing to cdk package
* Moves the coercing to the CDK package
