+++ src/cdk/coercion/boolean-property.spec.ts
@@ -1,7 +1,7 @@
 import {coerceBooleanProperty} from './boolean-property';
 
-
 describe('coerceBooleanProperty', () => {
+
   it('should coerce undefined to false', () => {
     expect(coerceBooleanProperty(undefined)).toBe(false);
   });
@@ -45,4 +45,5 @@ describe('coerceBooleanProperty', () => {
   it('should coerce an array to true', () => {
     expect(coerceBooleanProperty([])).toBe(true);
   });
+
 });
