+++ src/cdk/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+/*
+ * This file is not used to build this module. The real "index.js" file will be generated by the
+ * Angular Compiler CLI. This file is used inside of Google and is also used to simplify the
+ * SystemJS configuration
+ */
+export * from './public_api';
