+++ src/cdk/tsconfig-tests.json
@@ -11,6 +11,7 @@
     "experimentalDecorators": true
   },
   "include": [
-    "**/*.spec.ts"
+    "**/*.spec.ts",
+    "index.ts"
   ]
 }
