+++ src/lib/core/common-behaviors/disabled.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {coerceBooleanProperty} from '../coercion/boolean-property';
+import {coerceBooleanProperty} from '@angular/cdk';
 import {Constructor} from './constructor';
 
 /** @docs-private */
