+++ src/lib/core/option/option.ts
@@ -18,7 +18,7 @@ import {
   Optional,
 } from '@angular/core';
 import {ENTER, SPACE} from '../keyboard/keycodes';
-import {coerceBooleanProperty} from '../coercion/boolean-property';
+import {coerceBooleanProperty} from '@angular/cdk';
 import {MATERIAL_COMPATIBILITY_MODE} from '../../core/compatibility/compatibility';
 import {MdOptgroup} from './optgroup';
